pyproject.toml
[project]
name = "ses-partner-hub"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4",
    "lxml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest"]
